+++ components/site/patterns/Create Patterns/data.js
@@ -1,0 +1,24 @@
+// Comment the other and uncomment this to view/edit the empty state
+// export const serviceData = [];
+export const serviceData = [
+  {
+    name: 'Montanõ - Valencia',
+    sid: 'SM0yc4mxi6cn4z13bte7qmflc2drc85mlp',
+    created: 'Tue Jan 12 2021 15:42:05 GMT-0700 (Mountain Standard Time)',
+  },
+  {
+    name: 'Mante Inc',
+    sid: 'SMl29llgoihx286uhxfb0yc5n0sg391x5n',
+    created: 'Tue Jan 12 2021 15:42:05 GMT-0700 (Mountain Standard Time)',
+  },
+  {
+    name: 'Mante Inc',
+    sid: 'SMxarke3v30fv17hauqn86a7nhgm3b5d87',
+    created: 'Tue Jan 12 2021 15:42:05 GMT-0700 (Mountain Standard Time)',
+  },
+  {
+    name: 'Kihn - Ondricka',
+    sid: 'US0yc4mxi6cn4z13bte7qmflc2drc85mlp',
+    created: 'Tue Jan 12 2021 15:42:05 GMT-0700 (Mountain Standard Time)',
+  },
+];
