+++ package.json
@@ -21,7 +21,8 @@
     "prop-types": "15.7.2",
     "react": "16.14.0",
     "react-dom": "16.14.0",
-    "swr": "^0.4.0"
+    "swr": "^0.4.0",
+    "react-hook-form": "6.14.1"
   },
   "devDependencies": {
     "eslint": "7.17.0",
